flask-code/requirements.txt
Flask-Session
flask-swagger-ui
celery
flask_restful
boto3
requests
xarray
numpy
matplotlib
distributed
netcdf4
GitPython
GitHub-Flask
Flask-SQLAlchemy